Update tm351_22j.yml
I can't get permissions to work; the container runs as $NB_USER but it seems there's other issues running the workflow? I have no idea. Testing as root is not ideal because it maybe that there are permissions issues that testing would reveal if we were not running as root.